+++ extensions/firestore-stripe-payments-5056.env
@@ -1,0 +1,10 @@
+ALLOWED_EVENT_TYPES=com.stripe.v1.product.created
+CUSTOMERS_COLLECTION=customers
+DELETE_STRIPE_CUSTOMERS=Do not delete
+EVENTARC_CHANNEL=projects/pavelj-extensions1/locations/europe-west4/channels/firebase
+LOCATION=us-central1
+PRODUCTS_COLLECTION=products
+STRIPE_API_KEY=projects/233894764015/secrets/ext-firestore-stripe-payments-5056-STRIPE_API_KEY/versions/1
+STRIPE_CONFIG_COLLECTION=configuration
+SYNC_USERS_ON_CREATE=Do not sync
+TEST_BUCKET=pavelj-extensions1.appspot.com
