+++ extensions/firestore-stripe-payments-autopush1.env
@@ -1,0 +1,8 @@
+CUSTOMERS_COLLECTION=customers
+DELETE_STRIPE_CUSTOMERS=Do not delete
+LOCATION=us-central1
+PRODUCTS_COLLECTION=products
+STRIPE_API_KEY=projects/233894764015/secrets/ext-firestore-stripe-payments-autopush1-STRIPE_API_KEY/versions/1
+STRIPE_CONFIG_COLLECTION=configuration
+SYNC_USERS_ON_CREATE=Do not sync
+TEST_BUCKET=pavelj-extensions1.appspot.com
